+++ src/core/managers/internal/ToastManager/types/toast.types.ts
@@ -46,8 +46,10 @@ export interface ITransactionToast {
 }
 
 export enum ToastEventsEnum {
+  OPEN_NOTIFICATIONS_FEED = 'OPEN_NOTIFICATIONS_FEED',
   CLOSE_TOAST = 'CLOSE_TOAST',
+  // Event to update the transaction toast data
   TRANSACTION_TOAST_DATA_UPDATE = 'TRANSACTION_TOAST_DATA_UPDATE',
-  CUSTOM_TOAST_DATA_UPDATE = 'CUSTOM_TOAST_DATA_UPDATE',
-  VIEW_ALL = 'VIEW_ALL'
+  // Event to update the custom toast data
+  CUSTOM_TOAST_DATA_UPDATE = 'CUSTOM_TOAST_DATA_UPDATE'
 }