+++ src/utils/transactions/getInterpretedTransaction/helpers/getTokenFromData.ts
@@ -2,7 +2,7 @@ import { Address } from '@multiversx/sdk-core';
 import BigNumber from 'bignumber.js';
 import { TransactionTypesEnum } from 'types/enums.types';
 import { decodePart } from 'utils/decoders/decodePart';
-import { addressIsValid } from '../validation';
+import { addressIsValid } from 'utils/validation/addressIsValid';
 
 const noData = {
   tokenId: '',